Moved propose_random_flip into proposals.py
